Small changes for ergonimics and perf
Since the range of indices is always an actual range,
the return type of the `indices` function was changed
from `Vec<usize>` to a `Range<usize>` --- the user can
trivially create the corresponding `Vec` from this if they wish.

Added a function to return a shared refernece to an underlying
name (`&str`) rather than a clone of the name (`String`) which
is kind of a micro optimization, but could avoid unnecessary
allocations. The original function remains, however, in case
the user wants an owned copy (one could also consider returing
a CoW String).